[minor] "brute-force" refinement proof for size syscall.
I had some trouble proving an invariant of the calling convention
of the memory safety symbolic machine: when a syscall is entered, ra should be
a pointer a prescribed by the abstract machine. Not sure how to prove that this
is an invariant of the symbolic machine with our set of rules, though.