[minor] Fix build.
For some random reason, an unrelated change in the code base caused
typeclass inference to fail in a place where it was working just fine
under Coq 8.4pl4, but not under 8.4pl3. What is even more strange about
it, is that a similar inference had to be done right before and was
accepted, even on 8.4pl4. This could be worth investigating...